requirements.txt
ansicolors==1.1.8
cursor==1.2.0
